Create PayPalCheckoutRequest class
